requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
SQLAlchemy==2.0.31
Werkzeug==3.0.3
inference-sdk
python-dotenv
pandas
numpy
Pillow
gunicorn==21.2.0
